add stripe models

+++ app/models/account_type.rb
@@ -3,7 +3,7 @@ class AccountType
 
   field :name,       type: String
   field :data_limit, type: Integer
-  field :price,      type: Integer
+  field :price,      type: Float
 
   has_many :accounts
 end
